refactor(Input): make TextArea style tree shakeable
